refactor music.py, remove 0iq code 4 mega optimise
