+++ .gitignore
@@ -3,7 +3,7 @@
 frontend/.next
 frontend/node_modules
 frontend/.env
-
+bot/session.json
 
 # ------
