--- bot/requirements.txt
@@ -1,6 +1,0 @@
-quart 
-quart_cors
-quart-discord 
-discord-ext-ipc 
-discord
-Wavelink==1.3.5
